requirements/common.txt
Django==1.8.5
Pillow==2.8.1
psycopg2==2.5.2
wsgiref==0.1.2
djangorestframework==3.3.0
django-filter==0.12.0
python-dateutil==2.2
pytz==2013.9
django-crispy-forms==1.4.0
mandrill
django-rest-swagger==0.2.9
apns==1.1.2
requests==2.3.0
djrill==1.3.0
django-import-export==0.2.8
django-timezone-field==1.2.0
celery==3.1.17
django-solo==1.1.0
boto==2.39.0
django-extensions==1.5.2
django-guardian==1.3
dj-database-url==0.2.2
django-phonenumber-field==0.7.2
django-ses==0.7.1
sendgrid-django==2.0.0
